Add ability to set DebugScreen on specific instance
Spark has the `ignite` capability which allows the user to start more
than a single Service. It would be useful to attach debug screens to the
specific instances.

Ideally we should be able to re-use all the same methods and pass around
by default `getInstance` however access to the default instance remains
private in the Spark framework.